+++ CHANGELOG.md
@@ -6,3 +6,4 @@ ISSUES FIXED:
 * [49](https://github.com/perfectsense/gyro-azure-provider/issues/49): Public Ip Address Not Set at Startup for Network Interface.
 * [51](https://github.com/perfectsense/gyro-azure-provider/issues/51): Implement Lifecycle for Azure Storage.
 * [56](https://github.com/perfectsense/gyro-azure-provider/issues/56): Implement Azure managed Identities.
+* [62](https://github.com/perfectsense/gyro-azure-provider/issues/62): Add support to manage System Managed Service Identity for VM.